\name{AffyDistance-package}
\alias{AffyDistance-package}
\alias{AffyDistance}
\docType{package}
\title{
Affy support functions
}
\description{
A set of affy support functions and library calls.
}
\details{
\tabular{ll}{
Package: \tab AffyDistance\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-17\cr
License: \tab LGPL >= 2.0 \cr
Depends: \tab limma, gcrma, MASS, gamlss, stats4 \cr
LazyLoad: \tab yes\cr
}
Start by loading libraries, then transform the data, finally get the distance.
}
\author{
Corey M. Hudson
Maintainer: Corey M. Hudson <coreymhudson@gmail.com>
}
\references{
}
\keyword{ affy }
\seealso{
}
\examples{
}
