\name{loadAffyLibraries}
\alias{loadAffyLibraries}
\title{
Function to load necessary libraries
}
\description{
Function to load all the libraries that this package needs.
}
\usage{
loadAffyLibraries()
}
\details{
Function to load all the libraries that this package needs.
}
\value{
None
}
\references{
}
\author{
Corey M. Hudson <coreymhudson>
}
\note{
}


\examples{
function(){
	library(limma)
	library(gcrma)
	library(MASS)
	library(gamlss)
	library(stats4)
  }
}
\keyword{ libraries }
